Screen off animation [1/2]
Fade (default), CRT, and Scale

Credit and respect to xplodwild for paving the way
back in the KitKat days!

History of the ElectronBeam class can be found here
https://github.com/DirtyUnicorns/android_frameworks_base/commits/kitkat/services/java/com/android/server/power/ElectronBeam.java

*** Changes for Android 10 by bigrushdog ***
Various updates needed for Surface related API in the ElectronBeam animation class

[AgentFabulous | POSP]
- Rebase and rewrite ElectronBeam class on top of ColorFade
- Add usage of class-common Transaction instance
- Add support for respecting wide-color and protected-content
- Fixup for new API

@neobuddy89:
Updated for Android 14.

Change-Id: I58d269d44c901a8c0471807e3cf05c2054205d28
Co-Authored-By: Kshitij Gupta <kshitijgm@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Adapt Screen off animation [1/2] to Android 14 QPR2

Co-authored-by: afterallafk <shivamatiet2001@gmail.com>