VolumeDialogImpl: Ignore external layout direction changes
To simplify the handling of expanded volume rows with the dialog on the
left and right side, we use the layout direction to ensure the first row
is the outmost one. However, when the system layout direction was set
again but did not change its actual value after initDialog, this setting
was overriden and since no config change was indicated (mConfigChanged),
the system layout direction was used when opening the dialog the next
time. This lead to the first row not being the outmost one. To solve
that issue, just ignore system layout direction changes.

Test:
1. Open volume dialog
2. Eject and insert SIM card
3. Enter PIN, wait a second until the locale is set
4. Open the dialog and verify the the main row is the outmost one.
Change-Id: I0b9f15972d0416647ba045b7776bed8e6250d57f
Signed-off-by: danielml <daniel@danielml.dev>
Signed-off-by: Arian <arian.kulmer@web.de>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>